requirements.txt
autogen-agentchat>=0.2.0
fastmcp>=0.4.1
mcp>=1.1.0
pandas>=2.0
numpy>=1.25
matplotlib>=3.8
markdown>=3.6
jinja2>=3.1
weasyprint>=61.2
pypdf>=5.0.0
python-dotenv>=1.0.1
PyPDF2>=3.0.1